Update gen5.js
Black 2 White 2 Challenge/Easy modes are broken. 

The stats dont scale with the level changes made in either mode. So if a mon is level 56 it has the stats of a 56 even if its scaled up to 60 or down to 52. The change in natures and IVs still affect stats, but now the calc represents the levels that have the correct stat distribution.